pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "wav-takes-to-aaf"
version = "0.1.0"
description = "Convert a folder of multitrack WAV takes into a single AAF for Logic Pro import."
requires-python = ">=3.9"
license = { file = "LICENSE" }
dependencies = [
    "pyaaf2>=1.7.1",
]

[project.scripts]
wav-takes-to-aaf = "wav_takes_to_aaf.cli:main"

[tool.setuptools.packages.find]
include = ["wav_takes_to_aaf*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
